requirements.txt
python-telegram-bot
fastapi
uvicorn["standard"]